Refactor tests for DFT loaders.
