+++ scripts/AbinsModules/__init__.py
@@ -48,3 +48,4 @@ from . import AbinsParameters
 from . import AbinsConstants
 from . import AbinsTestHelpers
 
+from .GeneralLoadDFTTester import GeneralLoadDFTTester
